Move internal packages to internal path
Refactoring layout to match other projects that I maintain.
