+++ internal/config/deprecated.go
