+++ internal/config/merge_complete_configs_test.go
@@ -24,7 +24,7 @@ import (
 	"testing"
 
 	"github.com/alexflint/go-arg"
-	"github.com/atc0005/elbow/logging"
+	"github.com/atc0005/elbow/internal/logging"
 	"github.com/sirupsen/logrus"
 )
 
@@ -42,7 +42,7 @@ func TestMergeConfigUsingCompleteConfigObjects(t *testing.T) {
 
 	// Validation will fail if this is all we do since the default config
 	// doesn't contain any Paths to process.
-	baseConfig := NewDefaultConfig("x.y.z")
+	baseConfig := NewDefaultConfig()
 
 	testPaths := []string{"/tmp/elbow/path1"}
 	baseConfig.Paths = testPaths
